requirements.txt
# Phase 0 — Core dependencies
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic>=2.7.1
python-dotenv==1.0.1

# Phase 1+ — RAG pipeline (uncomment as you progress)
sentence-transformers==3.0.0; python_version < "3.13"
faiss-cpu==1.8.0; python_version < "3.13"
# chromadb==0.5.0
# langchain==0.2.0
# langchain-community==0.2.0
